ctamap_spider/pyproject.toml
[tool.poetry]
name = "ctamp"
version = "0.1.0"
description = ""
authors = ["Andy <zhangruibin021@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
selenium = "^4.0.0"
python-dateutil = "^2.8.2"
numpy = "1.19.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
